# API-CONTRACT.yaml
openapi: 3.1.0
info:
  title: Verdict API
  version: "1.0.0"
  description: >
    Canonical API contract for the Verdict hackathon judging platform.
    Every endpoint here must be implemented exactly as named — no additional
    endpoints invented by either the frontend or backend build without updating
    this file first. Enum values are defined once in ENUMS.md and referenced
    here by name. All list responses are wrapped in {"data": [...], "meta": {...}};
    all single-resource responses are wrapped in {"data": {...}}; all errors are
    {"error": {"code": "...", "message": "...", "details": {...}?}}.
servers:
  - url: /api/v1
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: verdict_session
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, enum: [unauthenticated, forbidden, not_found, validation_error, deadline_passed, event_not_open, conflict, rate_limited, internal_error] }
            message: { type: string }
            details: { type: object }
          required: [code, message]
      required: [error]
...
    EvaluationScoreInput:
      type: object
      properties:
        criterion_id: { type: string, format: uuid }
        value: { type: integer }
      required: [criterion_id, value]
    Evaluation:
      type: object
      properties:
        id: { type: string, format: uuid }
        assignment_id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        judge_user_id: { type: string, format: uuid }
        reaction: { type: string, enum: [not_convinced, interesting, stands_out], nullable: true }
        final_preference: { type: string, enum: [no, maybe, yes], nullable: true }
        tags:
          type: array
          items: { type: string, enum: [great_execution, clever_idea, strong_impact, beautiful_ux, technically_impressive, very_original, needs_more_polish] }
        comment: { type: string, nullable: true }
        scores: { type: array, items: { $ref: '#/components/schemas/EvaluationScoreInput' } }
        status: { type: string, enum: [in_progress, submitted] }
        submitted_at: { type: string, format: date-time, nullable: true }
...
  parameters:
    eventSlug:
      name: event_slug
      in: path
      required: true
      schema: { type: string }
...
security:
  - sessionCookie: []
paths:
...
  /events/{event_slug}:
    parameters: [ { $ref: '#/components/parameters/eventSlug' } ]
    get:
      summary: Event detail (public)
      security: []
      responses:
        '200': { description: ok }
        '404': { description: not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    patch:
      summary: Update event settings (organizer/admin only)
      responses:
        '200': { description: ok }
        '403': { description: forbidden, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /events/{event_slug}/publish-results:
    parameters: [ { $ref: '#/components/parameters/eventSlug' } ]
    post:
      summary: Publish the latest (or a specified) normalization run. Gated per JUDGING.md §5/§7.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                run_id: { type: string, format: uuid, description: "defaults to latest run" }
                force: { type: boolean, default: false }
                reason: { type: string, description: "required if force=true; written to audit log" }
      responses:
        '200': { description: published }
        '409': { description: coverage below min_publish_coverage_pct and force not set, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: forbidden, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  # ---------------------------------------------------------------- TRACKS / PRIZES
...
  /events/{event_slug}/judge/feed:
    parameters: [ { $ref: '#/components/parameters/eventSlug' } ]
    get:
      summary: Next unfinished assignment for the calling judge, with full project + rubric + progress counters, or null if queue empty
      responses: { '200': { description: ok }, '403': { description: caller is not a judge on this event } }
  /assignments/{assignment_id}/evaluation:
    parameters: [ { name: assignment_id, in: path, required: true, schema: { type: string, format: uuid } } ]
    post:
      summary: Create/update a draft evaluation (autosave — safe to call repeatedly, status stays in_progress)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reaction: { type: string, enum: [not_convinced, interesting, stands_out] }
                final_preference: { type: string, enum: [no, maybe, yes] }
                tags: { type: array, items: { type: string } }
                comment: { type: string }
                scores: { type: array, items: { $ref: '#/components/schemas/EvaluationScoreInput' } }
      responses:
        '200': { description: ok, content: { application/json: { schema: { type: object, properties: { data: { $ref: '#/components/schemas/Evaluation' } } } } } }
        '403': { description: assignment not owned by caller }
  /assignments/{assignment_id}/evaluation/submit:
    parameters: [ { name: assignment_id, in: path, required: true, schema: { type: string, format: uuid } } ]
    post:
      summary: Finalize the evaluation — requires all rubric criteria answered; sets status=submitted, locks further edits by the judge
      responses:
        '200': { description: ok }
        '422': { description: rubric incomplete, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: assignment not owned by caller }
  /events/{event_slug}/evaluations/mine:
    parameters: [ { $ref: '#/components/parameters/eventSlug' } ]
    get:
      summary: >
        Caller's own submitted + in-progress evaluations for this event. This is the
        .dogfood.toml `routes.judge_scores` target.
      responses: { '200': { description: ok }, '403': { description: caller has no judge/organizer/admin role } }
  /events/{event_slug}/judges/{judge_user_id}/evaluations:
    parameters:
      - { $ref: '#/components/parameters/eventSlug' }
      - { name: judge_user_id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      summary: >
        A specific judge's evaluations. This is the .dogfood.toml `routes.peer_scores` target.
        Allowed only if caller.id == judge_user_id, or caller is organizer/admin of this event —
        every other caller (including other judges) gets 403. This is the exact check the T2
        acceptance test targets; implement it once in the judging module's ownership check.
      responses:
        '200': { description: ok }
        '403': { description: forbidden — caller is neither the judge in question nor an organizer/admin, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  # ---------------------------------------------------------------- PAIRWISE (bonus)
...